requirements.txt
pyfred-cli @ git+https://github.com/acourtneybrown/pyfred-cli@main

humanize==4.12.3
requests==2.32.4
